Remove self signed certificate from production.
