+++ examples/client_sync.py
@@ -108,7 +108,8 @@ def setup_sync_client(description=None, cmdline=None):
             timeout=args.timeout,
             #    retries=3,
             # TLS setup parameters
-            sslctx=modbusClient.ModbusTlsClient.generate_ssl(
+            sslctx=helper.generate_ssl(
+                False,
                 certfile=helper.get_certificate("crt"),
                 keyfile=helper.get_certificate("key"),
                 #    password=None,
