+++ pymodbus/server/server.py
@@ -92,9 +92,6 @@ class ModbusTlsServer(ModbusTcpServer):
         identity: ModbusDeviceIdentification | None = None,
         address: tuple[str, int] = ("", 502),
         sslctx=None,
-        certfile=None,
-        keyfile=None,
-        password=None,
         ignore_missing_devices=False,
         broadcast_enable=False,
         trace_packet: Callable[[bool, bytes], bytes] | None = None,
@@ -111,11 +108,7 @@ class ModbusTlsServer(ModbusTcpServer):
         :param framer: The framer strategy to use
         :param identity: An optional identify structure
         :param address: An optional (interface, port) to bind to.
-        :param sslctx: The SSLContext to use for TLS (default None and auto
-                       create)
-        :param certfile: The cert file path for TLS (used if sslctx is None)
-        :param keyfile: The key file path for TLS (used if sslctx is None)
-        :param password: The password for for decrypting the private key file
+        :param sslctx: The SSLContext to use for TLS
         :param ignore_missing_devices: True to not send errors on a  missing device
         :param broadcast_enable: True to treat dev_id 0 as broadcast address,
                         False to treat 0 as any other dev_id
@@ -130,9 +123,7 @@ class ModbusTlsServer(ModbusTcpServer):
             reconnect_delay=0.0,
             reconnect_delay_max=0.0,
             timeout_connect=0.0,
-            sslctx=CommParams.generate_ssl(
-                True, certfile, keyfile, password, sslctx=sslctx
-            ),
+            sslctx=sslctx,
         )
         super().__init__(
             context,
