+++ pymodbus/transport/transport.py
@@ -100,34 +100,6 @@ class CommParams:
     stopbits: int = -1
     handle_local_echo: bool = False
 
-    @classmethod
-    def generate_ssl(
-        cls,
-        is_server: bool,
-        certfile: str | None = None,
-        keyfile: str | None = None,
-        password: str | None = None,
-        sslctx: ssl.SSLContext | None = None,
-    ) -> ssl.SSLContext:
-        """Generate sslctx from cert/key/password.
-
-        MODBUS/TCP Security Protocol Specification demands TLSv2 at least
-        """
-        if sslctx:
-            return sslctx
-        new_sslctx = ssl.SSLContext(
-            ssl.PROTOCOL_TLS_SERVER if is_server else ssl.PROTOCOL_TLS_CLIENT
-        )
-        new_sslctx.check_hostname = False
-        new_sslctx.verify_mode = ssl.CERT_NONE
-        new_sslctx.minimum_version = ssl.TLSVersion.TLSv1_2
-        new_sslctx.maximum_version = ssl.TLSVersion.TLSv1_3
-        if certfile:
-            new_sslctx.load_cert_chain(
-                certfile=certfile, keyfile=keyfile, password=password
-            )
-        return new_sslctx
-
     def copy(self: CommParams) -> CommParams:
         """Create a copy."""
         return copy.copy(self)
