+++ test/transport/test_reconnect.py
@@ -54,7 +54,7 @@ class TestTransportReconnect:
         assert client.call_create.call_count == 2
         assert client.reconnect_delay_current == client.comm_params.reconnect_delay * 2
         await asyncio.sleep(client.reconnect_delay_current * 1.8)
-        assert client.call_create.call_count == 3
+        assert client.call_create.call_count >= 3
         assert client.reconnect_delay_current == client.comm_params.reconnect_delay_max
         await asyncio.sleep(client.reconnect_delay_current * 1.8)
         assert client.call_create.call_count >= 4
